+++ src/commands/registerCommands.ts
@@ -1,5 +1,5 @@
 import { commands, Disposable, ExtensionContext } from 'vscode'
-import { run } from './run'
+import { run, showOutput } from './run'
 import { install, update } from './installation'
 import { openExample } from './openExample'
 
@@ -9,6 +9,7 @@ export function registerCommands (context: ExtensionContext): Disposable[] {
     commands.registerCommand('vscode-esbmc.verify.function', async (overrides, commentFlags) => run(overrides, commentFlags)),
     commands.registerCommand('vscode-esbmc.install', async () => install(context)),
     commands.registerCommand('vscode-esbmc.update', async () => update(context)),
+    commands.registerCommand('vscode-esbmc.showOutput', () => showOutput()),
     commands.registerCommand('vscode-esbmc.openExample', async () => openExample(context))
   ]
 }